+++ Plane_frame_pulse_response_simulation/src/geometry_store/pulse_result_objects/pulse_analysis_result_store.h
@@ -13,7 +13,7 @@ public:
 
 	pulse_analysis_result_store();
 	~pulse_analysis_result_store();
-	void set_analysis_setting(const int& time_step_count, const int& time_interval, const double& total_simulation_time);
+	void set_analysis_setting(const int& time_step_count, const double& time_interval, const double& total_simulation_time);
 	void clear_results();
 
 private:
